Multi-Label Class Sequence Classification
Study of Parameter Efficient Fine-Tuning using Low Rank Adaptaion (LoRa) method with a custom/public dataset on HuggingFace. I selected the BERT model and fine-tuned the base model for a Multi-Class (6) labeled dataset (https://huggingface.co/datasets/dair-ai/emotion). This practice serves as a helpful introduction for newcomers looking to customize the base model for a particular task.Fine-tuning BERT using the LoRa method on a custom dataset not only improves model performance but also reduces computational resources required for training. This approach can be easily extended to other NLP tasks and datasets, making it a valuable technique for researchers and practitioners in the field.